# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: game
    description: Kitaba game operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /game/new:
    post:
      operationId: newGame
      tags: [game]
      summary: Start a new game session
      description: Creates a new game session with the initial world state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewGameRequest"
      responses:
        "200":
          description: New game session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /game/action:
    post:
      operationId: submitAction
      tags: [game]
      summary: Submit a player action
      description: Processes a player action and returns narrative + updated world state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionRequest"
      responses:
        "200":
          description: Action processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /game/save:
    post:
      operationId: saveGame
      tags: [game]
      summary: Save the current game
      description: Saves the current session state with an optional name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveGameRequest"
      responses:
        "200":
          description: Game saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveGameResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /game/load:
    post:
      operationId: loadGame
      tags: [game]
      summary: Load a saved game
      description: Loads a previously saved game session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadGameRequest"
      responses:
        "200":
          description: Game loaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /game/saves:
    get:
      operationId: listSaves
      tags: [game]
      summary: List available saves
      description: Returns a list of saved game slots
      responses:
        "200":
          description: List of saves
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested game resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: World version changed before the action could be committed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NewGameRequest:
      type: object
      properties:
        playerName:
          type: string
          description: The player character's name
      required:
        - playerName

    ActionRequest:
      type: object
      properties:
        sessionId:
          type: string
        playerInput:
          type: string
          description: Free-form text describing what the player does
      required:
        - sessionId
        - playerInput

    SaveGameRequest:
      type: object
      properties:
        sessionId:
          type: string
        saveName:
          type: string
      required:
        - sessionId
        - saveName

    LoadGameRequest:
      type: object
      properties:
        saveId:
          type: string
      required:
        - saveId

    CharacterStatus:
      type: object
      properties:
        name:
          type: string
        locationName:
          type: string
        worldDate:
          type: string
        hunger:
          type: number
          description: 0 (starving) to 100 (full)
        fatigue:
          type: number
          description: 0 (exhausted) to 100 (rested)
        health:
          type: number
          description: 0 (dead) to 100 (perfect)
      required:
        - name
        - locationName
        - worldDate
        - hunger
        - fatigue
        - health

    NarrativeEntry:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [player, narrator, system]
        text:
          type: string
        timestamp:
          type: string
      required:
        - id
        - type
        - text
        - timestamp

    GameSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        characterStatus:
          $ref: "#/components/schemas/CharacterStatus"
        narrativeHistory:
          type: array
          items:
            $ref: "#/components/schemas/NarrativeEntry"
        introText:
          type: string
      required:
        - sessionId
        - characterStatus
        - narrativeHistory

    ActionResponse:
      type: object
      properties:
        sessionId:
          type: string
        narrativeEntry:
          $ref: "#/components/schemas/NarrativeEntry"
        characterStatus:
          $ref: "#/components/schemas/CharacterStatus"
        worldVersion:
          type: integer
          minimum: 0
      required:
        - sessionId
        - narrativeEntry
        - characterStatus
        - worldVersion

    SaveGameResponse:
      type: object
      properties:
        saveId:
          type: string
        saveName:
          type: string
        savedAt:
          type: string
      required:
        - saveId
        - saveName
        - savedAt

    SaveInfo:
      type: object
      properties:
        saveId:
          type: string
        saveName:
          type: string
        characterName:
          type: string
        savedAt:
          type: string
        worldVersion:
          type: integer
          minimum: 0
      required:
        - saveId
        - saveName
        - characterName
        - savedAt
        - worldVersion

    SaveListResponse:
      type: object
      properties:
        saves:
          type: array
          items:
            $ref: "#/components/schemas/SaveInfo"
      required:
        - saves
